+++ RELEASE_NOTES.md
@@ -26,6 +26,7 @@
 * Update the SDK dependency to `1.0.0rc901`.
 * Change `edit_uri` default branch to v0.x.x in mkdocs.yml.
 * Added a new default option `asyncio_default_fixture_loop_scope = "function"` for `pytest-asyncio` as not providing a value is deprecated.
+* The migration script is now written in Python, so it should be (hopefully) more compatible with different OSes.
 
 ## Bug Fixes